+++ integer-types/library/Integer/Sign.hs
@@ -8,13 +8,18 @@ module Integer.Sign
 import Essentials
 
 import Prelude (seq)
+import qualified Prelude as Enum (Enum (..))
+import Data.Hashable (Hashable (hashWithSalt))
 
 import qualified Control.DeepSeq as DeepSeq
 
 data Sign = MinusSign | PlusSign
-    deriving (Eq, Ord, Show)
+    deriving stock (Eq, Ord, Show, Enum, Bounded)
 
 instance DeepSeq.NFData Sign where rnf x = seq x ()
+
+instance Hashable Sign where
+    hashWithSalt salt x = salt `hashWithSalt` (Enum.fromEnum x)
 
 negate :: Sign -> Sign
 negate PlusSign  = MinusSign